requirements.txt
absl-py==2.3.1
catboost==1.2.8
clear==2.0.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
graphviz==0.21
immutabledict==4.2.2
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
matplotlib==3.10.7
narwhals==2.11.0
numpy==2.3.4
ortools==9.14.6206
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
protobuf==6.31.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
xgboost==3.1.1
